#587: added games
