--- packages/gdl-frontend/components/BookDetailsPage/index.js
@@ -1,10 +1,0 @@
-// @flow
-/**
- * Part of GDL gdl-frontend.
- * Copyright (C) 2018 GDL
- *
- * See LICENSE
- */
-
-export { default as BookJsonLd } from './BookJsonLd';
-export { default as Metadata } from './Metadata';
