+++ packages/gdl-frontend/components/DetailsPage/__tests__/BookJsonLd.test.js
