+++ packages/gdl-frontend/components/DetailsPage/index.js
@@ -1,0 +1,14 @@
+// @flow
+/**
+ * Part of GDL gdl-frontend.
+ * Copyright (C) 2018 GDL
+ *
+ * See LICENSE
+ */
+
+export { default as BookJsonLd } from './BookJsonLd';
+export { default as Metadata } from './Metadata';
+export { default as Grid } from './Grid';
+export { default as GridItem } from './GridItem';
+export { default as Divider } from './Divider';
+export { default as ShareButton } from './ShareButton';
