+++ packages/gdl-frontend/routes.js
@@ -48,8 +48,6 @@ routes.add('book', `/${langParam}/books/details/:id(\\d+)`, 'books/_book');
 // Game details page
 routes.add('game', `/${langParam}/games/details/:id`, 'games/_game');
 
-routes.add('browseGames', `/${langParam}/games/browse`, 'games/browse');
-
 // Read book
 routes.add(
   'read',
